// frontend/src/api/generated/schemas/packageItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SecureShipBackend
 * BFF for SecureShip client application
 * OpenAPI spec version: 1.0.0
 */

export interface PackageItem {
  id: string;
  shipment_id: string;
  description: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  weight_kg: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  declared_value: string;
  shipment_tracking_number?: string | null;
}